Converts any number in base 10 to another base
The program receives two inputs:
First, a number in base 10 and then the base to which the number is to be converted to.
Then, it takes the modulus of the number and the base and appends it to an empty list. I wrapped it in a loop which takes the following procedures:
1. Number (mod) base and appends it to a list.
2. Number ÷ base. The result becomes the new number and it goes back to step 1.
The loop cycles on until the number becomes 0. The loop breaks, the list is reversed and printed.