A Car Game Engine
This was an exercise by Mosh Hamedani. The program receives start and stop prompts to start or stop the car respectively.  The subsequent assignment was to advance the program to "yell" that the car has already started or stopped, when the user inputs start or stop successively. 

What I included was an extra if nest to output that the car was never started if the user inputs stop before start as a first command.