An Emoji Converter Program
This is an emoji converter exercise from Mosh. The program receives an input and matches the input to the values in the dictionary. If it matches, it prints the definition of the word, otherwise, it prints out the input as it is. I don't see how I can advance it since the user must have to know the keywords in the dictionary otherwise, it is just an input-print program.