A Guessing Game
The user has three chances to guess a secret number. Three things I advanced in this game are:
Firstly, I made the secret number a random selection. It is not an assigned variable which makes it interesting as both programmer and the user are ignorant of what the value is. They would both have to GUESS! 

Because of the randomness, I added hint programs to help with the guessing. I wrote elif statements to output too hot or too cold to indicate that the guess was greater than or less than the secret number.

Finally, I enveloped the code in a try-except block to catch ValueError. Initially, I did not like that when an invalid input is made, the program does not subtract from the available guesses. But I reckoned that the program should forgive an invalid input that was made mistakenly. Therefore, making an invalid input (like a string) does not count as a guess. I set up the try block inside the while loop so, the game continues after the wrong input prompt.