A Debit Card Class
I created a debit card class and gave it withdraw, transfer and check balance functions.

In the future, I want to make this code a little more elegant by not having to put explicitly type pin input for every function. The functions would demand the input.

I could declare a function that requests for the pin and call it inside the debit card functions. But I believe there's a way to declare inputs and call them like parameters to the functions.

When I learn that, I'll recommit this file.