A Debit Card Class (Updated)
This time, rather than embedding input statements into every function, I declared individual functions and passed under it, the input statements for all the needed inputs. (pin, amount, etc.). So, whenever I need them (for withdrawal or transfer), I would just call the function of the required input inside the function. 

Also, I added a layer of authentication to the pin and recipient account number. Using if statements, I regulated the pin to always be four digits and the recipient account number to always be 10. Anything below or above is invalid and I wrapped it in a while loop so, the program does not continue until the right data is passed. 

Also, I added another function, Check Balance.