Eric's $50,000 giveaway game
This is a guessing game but with much more extras. It is the syntax of the Bagels game by Ai Sweigart. I changed the storyline and the layout of the code. 

One thing I love is how I arranged the code slightly abnormally. I named the statement variables and called those names when I wanted to print those statements. This made the functions shorter and easier to glance through.